+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
 * @mmcc added a VERSION key to the videojs object ([view](https://github.com/videojs/video.js/pull/1798))
 * @mmcc fixed an issue with text track hiding introduced in #1681 ([view](https://github.com/videojs/video.js/pull/1804))
 * Export video.js as a named AMD module ([view](https://github.com/videojs/video.js/pull/1844))
+* Hide the poster when play fires ([view](https://github.com/videojs/video.js/pull/1834))
 
 --------------------